Create Stopwatch web application code Task 02
To build a stopwatch web application ,used  HTML, CSS, and JavaScript.HTML is used for structure the elements of the application. By implementing functions for starting, pausing, and resetting the stopwatch, as well as tracking and displaying lap times, users can accurately measure and record time intervals. With these technologies and functionalities, you can create an interactive and user-friendly stopwatch web application